+++ pysys-examples/internal/testcases/PySys_internal_129/run.py
@@ -21,7 +21,7 @@ class PySysTest(BaseTest):
 		# this should NOT stop us deleting this output dir
 		self.write_text(self.mkdir('testroot/__pysys_output_archives_lowTotalLimit/test_output')+'/myfile.txt', 'xxx')
 		
-		runPySys(self, 'pysys', ['run', '-o', self.output+'/test_output',
+		runPySys(self, 'pysys', ['run', '-o', 'test_output',
 			'--record', '--threads', '6', '-c', '10'
 		], workingDir='testroot', expectedExitStatus='!=0')